oil-pump-monitor/pyproject.toml
[project]
name = "oil-pump-monitor"
version = "0.1.0"
description = "Oil Pump Vibration Monitoring - North Dakota Bakken Formation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "aiohttp>=3.9.0",
    "databricks-sdk>=0.30.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
]
